Rename read_scheduler to download_scheduler
