+++ src/controller/mod.rs
@@ -1,4 +1,4 @@
-pub mod read_scheduler;
+pub mod download_scheduler;
 pub mod stream;
 pub mod task_manager;
 pub mod timeouts;
